changed the delete button to include a warning message with it confirming the user wants to dele them. Added a menu for version number and a messagebox for the version number

+++ WindowsFormsApplication1/SearchContactForm.cs
@@ -75,7 +75,15 @@ namespace WindowsFormsApplication1
 
             try
             {
-                ContactDB.DeleteContact(conID);
+                if (MessageBox.Show("Deleting " + delContact.ContactFirstName + " " + delContact.ContactLastName + " from the database is an action that can NOT be undone!\n\nAre you sure you want to delete this contact?", "CONFIRM DELETE", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation) == DialogResult.Yes)
+                {
+                    ContactDB.DeleteContact(conID);
+                }
+                else
+                {
+                }
+
+
 
                 // Refreshes the gridview
                 contactListing = ContactDB.GetAllContacts();
